pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "building-intelligent-agents"
version = "0.1.0"
description = "Code for Building intelligent agents using Google ADK"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.0.0",
    "langchain-google-genai",
    "pydantic>=2.0, <3.0.0",
    "email-validator>=1.1",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "crewai-tools>=0.45.0",
    "langchain-community>=0.3.24",
    "langgraph>=0.4.7",
    "duckduckgo-search>=8.0.2",
    "docker>=7.1.0",
    #"google-cloud-aiplatform>=1.95.1", # Uncomment if using Google Cloud Vertex AI
    "anthropic>=0.52.0",
    "litellm>=1.68.0",
    "google-adk[eval]"
]

[tool.setuptools]
package-dir = {"" = "src"} # Tells setuptools that packages are under the "src" directory

[tool.setuptools.packages.find]
where = ["src"]  # Location of packages relative to pyproject.toml
include = ["building_intelligent_agents", "building_intelligent_agents.*"] # Packages to include